[feat] 인자 타입 변경

+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/ResumeMyCertScreen.kt
@@ -79,7 +79,7 @@ fun ResumeMyCertRoute(
 fun ResumeMyCertScreen(
     certifications: List<CertificationData>,
     showDialog: Boolean,
-    onDeleteClick: () -> Unit,
+    onDeleteClick: (Long) -> Unit,
     onDialogConfirm: () -> Unit,
     onDialogDismiss: () -> Unit,
     modifier: Modifier = Modifier
